+++ core/design-system/src/main/java/com/twix/designsystem/components/button/model/LoginTypeUiModel.kt
@@ -1,7 +1,7 @@
 package com.twix.designsystem.components.button.model
 
 import androidx.compose.ui.graphics.Color
-import com.twix.domain.login.LoginType
+import com.twix.domain.model.enums.LoginType
 
 data class LoginTypeUiModel(
     val type: LoginType,
