+++ domain/src/main/java/com/twix/domain/login/LoginProvider.kt
@@ -1,5 +1,7 @@
 package com.twix.domain.login
 
+import com.twix.domain.model.enums.LoginType
+
 interface LoginProvider {
     suspend fun login(): LoginResult
