+++ domain/src/main/java/com/twix/domain/login/LoginResult.kt
@@ -1,5 +1,7 @@
 package com.twix.domain.login
 
+import com.twix.domain.model.enums.LoginType
+
 sealed interface LoginResult {
     data class Success(
         val idToken: String,
