+++ feature/login/src/main/java/com/twix/login/LoginProviderFactory.kt
@@ -1,7 +1,7 @@
 package com.twix.login
 
 import com.twix.domain.login.LoginProvider
-import com.twix.domain.login.LoginType
+import com.twix.domain.model.enums.LoginType
 
 class LoginProviderFactory(
     private val providers: Map<LoginType, LoginProvider>,
